fix leaky assert in nbdcache
The assert in nbdcache checks that a valid response from the cache does not have
any exceptions, as it is the job of the pipeline to detect asserts and supply
the kill signal to the cache. However, the assert can be erroneously triggered
by a load that misses in the cache returning and being compared against spurious
exception signals.

This commit checks that the tag of the response matches the tag corresponding to
the exceptions and also checks that the excepting instruction was valid.